Added new datalayer <-> views translation layer
This file is responsible for composing elements of the datalayer that
the views need access to. It should be the only part of the application
that has some notion of what the datalayer expects and what the view
returns. If there is a rough patch to be smoothed over, the buck stops
here. ~> (buck)